no message
git-svn-id: https://mosesdecoder.svn.sourceforge.net/svnroot/mosesdecoder@87 1f5c12ca-751b-0410-a591-d2e778427230

+++ trunk/moses/src/StaticData.h
@@ -51,7 +51,7 @@ protected:
 									// do it differently from old pharaoh
 									// -ve	= no limit on distortion
 									// 0		= no disortion (monotone in old pharaoh)
-	unsigned int                                      m_maxHypoStackSize; //hypothesis-stack size that triggers pruning
+	size_t                              m_maxHypoStackSize; //hypothesis-stack size that triggers pruning
 	size_t															m_nBestSize;
 	std::string													m_nBestFilePath, m_cachePath;
 	std::vector<std::string>						m_mySQLParam;
